던전 입장 추가

+++ Armor.cs
@@ -9,8 +9,8 @@ namespace TextRPG
 {
     internal class Armor: Item
     {
-        public double def { get; set; }
-        public Armor(string name, double def, string info, int price, bool isSetting) : base(name, info, price)
+        public int def { get; set; }
+        public Armor(string name, int def, string info, int price, bool isSetting) : base(name, info, price)
         {
             this.def = def;
             this.isSetting = isSetting;
