+++ Weapon.cs
@@ -8,8 +8,8 @@ namespace TextRPG
 {
     internal class Weapon: Item
     {
-        public double atk {  get; set; }
-        public Weapon(string name, double atk, string info, int price, bool isSetting): base(name, info, price)
+        public int atk {  get; set; }
+        public Weapon(string name, int atk, string info, int price, bool isSetting): base(name, info, price)
         {
             this.atk = atk;
             this.isSetting = isSetting;
